refactor: renaming root folder for project, changing location of traits and macro implementation
