+++ smithy4play/src/main/scala/de/innfactory/smithy4play/RoutingContext.scala
@@ -1,0 +1,10 @@
+package de.innfactory.smithy4play
+
+import play.api.mvc.{RawBuffer, Request}
+
+case class RoutingContext(map: Map[String, Seq[String]])
+
+object RoutingContext {
+  def fromRequest(request: Request[RawBuffer]): RoutingContext =
+    RoutingContext(request.headers.toMap)
+}
