+++ smithy4play/src/main/scala/de/innfactory/smithy4play/RoutingErrorResponse.scala
